improved support for tags in frame names
added tags:
LOOP - next frame is the first in the animation
NOLOOP - next frame is null
JUMPTO X - jump to the given frame index
any name not as a known tag is interpreted as the name of the animation. the names and their frame index will be exported as defines in the header.